third_party/blink/web_tests/webmidi/permission.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../http/tests/resources/permissions-helper.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/media/midi/midi_service.mojom.js"></script>
<script src="../http/tests/webmidi/mock-midiservice.js"></script>
</head>
<body>
<script>
promise_test(async _ => {
  await PermissionsHelper.setPermission('midi', 'granted');
  await PermissionsHelper.setPermission('midi-sysex', 'granted');

  const access = await navigator.requestMIDIAccess({sysex: true});
  assert_true(access.sysexEnabled);
}, 'sysex permission grants are properly handled');

promise_test(async t => {
  await PermissionsHelper.setPermission('midi', 'granted');
  await PermissionsHelper.setPermission('midi-sysex', 'denied');

  return promise_rejects_dom(t, 'SecurityError',
      navigator.requestMIDIAccess({sysex: true}));
}, 'sysex permission rejections are properly handled');

</script>
</body>
</html>
